Fixed a bug that could prevent a Tree from being removed by Tree Feller if 2 or more players were using Tree Feller simultaneously
